+++ PassWard/Config/PluginConfig.cs
@@ -2,6 +2,8 @@
 
 using BepInEx.Configuration;
 
+using ComfyLib;
+
 using UnityEngine;
 
 public static class PluginConfig {
@@ -17,33 +19,49 @@ public static class PluginConfig {
 
   public static ConfigEntry<UserListSeperator> WardHoverTextUserListSeparator { get; private set; }
 
+  public enum UserListSorting {
+    Unsorted,
+    Alphabetically
+  }
+
+  public static ConfigEntry<UserListSorting> WardHoverTextUserListSorting { get; private set; }
+
   public static void BindConfig(ConfigFile config) {
     IsModEnabled =
-        config.Bind(
+        config.BindInOrder(
             "_Global",
             "isModEnabled",
             true,
             "Globally enable or disable this mod.");
 
     EnterPasswordKey =
-        config.Bind(
+        config.BindInOrder(
           "Hotkeys",
           "enterPasswordShortcut",
           new KeyboardShortcut(KeyCode.P, KeyCode.LeftShift),
           "Enter password into a passworded ward OR assign password to own ward.");
 
     RemovePasswordKey =
-        config.Bind(
+        config.BindInOrder(
           "Hotkeys",
           "removePasswordKey",
           new KeyboardShortcut(KeyCode.R, KeyCode.LeftShift),
           "Removes password on player's own passworded ward.");
 
     WardHoverTextUserListSeparator =
-        config.Bind(
+        config.BindInOrder(
             "HoverText",
             "userListSeparator",
             UserListSeperator.Newline,
             "Separator to use between player names for a ward's hover-text.");
+
+    WardHoverTextUserListSorting =
+        config.BindInOrder(
+            "HoverText",
+            "userListSorting",
+            UserListSorting.Unsorted,
+            "What sorting (if any) to use for player names for a ward's hover-text.");
+
+    WardHoverTextUserListSorting.OnSettingChanged(WardManager.ClearCachedPermittedPlayerNames);
   }
 }