+++ PassWard/Core/WardManager.cs
@@ -1,6 +1,10 @@
 ﻿namespace PassWard;
 
+using System;
+using System.Collections.Generic;
 using System.Text;
+
+using UnityEngine;
 
 using static PluginConfig;
 
@@ -98,6 +102,68 @@ public static class WardManager {
   public static void ShowMessage(string message) {
     if (MessageHud.instance) {
       MessageHud.instance.ShowMessage(MessageHud.MessageType.TopLeft, message);
+    }
+  }
+
+  static readonly List<string> _cachedPlayerNames = new(capacity: 100);
+
+  static ZDOID _lastZDOID = ZDOID.None;
+  static uint _lastDataRevision = uint.MaxValue;
+
+  public static void ClearCachedPermittedPlayerNames() {
+    _lastZDOID = ZDOID.None;
+    _lastDataRevision = uint.MaxValue;
+    _cachedPlayerNames.Clear();
+  }
+
+  public static List<string> GetCachedPermittedPlayerNames(ZDO zdo) {
+    if (_lastZDOID == zdo.m_uid && _lastDataRevision == zdo.DataRevision) {
+      return _cachedPlayerNames;
+    }
+
+    _lastZDOID = zdo.m_uid;
+    _lastDataRevision = zdo.DataRevision;
+    _cachedPlayerNames.Clear();
+
+    int permittedCount = Mathf.Max(zdo.GetInt(ZDOVars.s_permitted, 0), 0);
+
+    CachePuIdNameHashCodes(permittedCount);
+
+    for (int i = 0; i < permittedCount; i++) {
+      if (zdo.GetLong(_cachedPuIdHashCodes[i], 0L) == 0L) {
+        continue;
+      }
+
+      string playerName = zdo.GetString(_cachedPuNameHashCodes[i], string.Empty);
+
+      if (playerName.Length > 0) {
+        _cachedPlayerNames.Add(playerName);
+      }
+    }
+
+    if (WardHoverTextUserListSorting.Value == UserListSorting.Alphabetically) {
+      _cachedPlayerNames.Sort(StringComparer.OrdinalIgnoreCase);
+    }
+
+    return _cachedPlayerNames;
+  }
+
+  static int[] _cachedPuIdHashCodes = [];
+  static int[] _cachedPuNameHashCodes = [];
+
+  static void CachePuIdNameHashCodes(int capacity) {
+    int count = Mathf.Min(_cachedPuIdHashCodes.Length, _cachedPuNameHashCodes.Length);
+
+    if (capacity <= 0 || capacity < count) {
+      return;
+    }
+
+    Array.Resize(ref _cachedPuIdHashCodes, capacity);
+    Array.Resize(ref _cachedPuNameHashCodes, capacity);
+
+    for (int i = count; i < capacity; i++) {
+      _cachedPuIdHashCodes[i] = $"pu_id{i}".GetStableHashCode();
+      _cachedPuNameHashCodes[i] = $"pu_name{i}".GetStableHashCode();
     }
   }
 }