+++ PassWard/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 1.2.0
+
+  * Added new config-option `HoverText.userListSorting`.
+  * Added new logic to cache permitted-player names to improve performance.
+
 ### 1.1.0
 
   * Fixed for the `v0.220.3` patch.